Fix counts file containing invalid data
The dictionary created by the tag exploder is not merged
correctly with the count dictionary.
The reason for this is the addition of "from section" and "to section" tags.
This will result in new tags that do not match with the tags contained in
the count dictionary.
To solve this the logic to add the section info has now been implemented as
another `Exporter` / `ExporterFactory` that acts as a decorator to existing
exporters.