Fix exporting counts creating double entries with value count of zero
The `TagExploder` uses actual datetime to create the start and end timestamps.
The problem here is that the timestamps take the seconds into account when
creating a timeslot tag.
On the other hand, the tags associated with their respective counts omit the seconds and only take into account date, hour and minutes.
Thus leading to double entries during the merge of the two dicts.