Show info that offset of currently selected section and visualized offset are different
The 'Update Section Offset' button will be colored orange and enabled, if the visualisation
offset differs from the section offset. Otherwise, the button will be disabled and its color
will be set to default.
If multiple or no sections are selected, the button color  will be disabled and its color
will be set to default.