Give observers listening on track repository changes information on removed tracks
This will fix a track visualisation error where not all
tracks are visualised after the cutting of tracks due
to CachedPandasTrackProvider not being able to distinguish
between removed and added tracks. This caused a reset of the
cache although only a subset of tracks has been removed.
A subsequent add of new cut tracks to the cache results in
a dataframe of only cut tracks. The cache only retrieves
tracks from the track repository if the cache is empty.