Do not filter tracks with certain number of tracks when creating a new PandasTrackDataset
The filtering by the number of detections should be the sole
responsibility of the TrackParser. By using a cutting section
it is possible and valid to have tracks less detections than
the given limit used by the TrackParser. Thus, no
filtering should be applied when creating a new TrackDataset,
resulting in missing tracks.