Setup a RoadUserAssignmentRepository to store previously computed RUAs, they are neither counted nor tagged. Components using tagged assignments can now get stored assignments and then apply a tagger.
Moved RoadUserAssignment, EventPair and RoadUserAssignments to that file
RoadUserAssignmentRepository allows adding either lists[RoadUserAssignment] or RoadUserAssignments
and returns all computed assignments either as list or as RoadUserAssignments
To avoid dependency of RoadUserAssignments to Tagger, moved Tag function from RoadUserAssignments to Tagger

Added use case: CreateRoadUserAssignments. It will apply a given RoadUserAssigner to the events of the EventRepository. If event creation is enabled and the event repository is empty, CreateEvents use case is triggered first. When executing the use case, an overwrite flag can be set, specifying that the repository content should be cleared and recomputed.

Modified use case: GetRoadUserAssignments. GetRoadUserAssignments no longer applies assignments itself but rather gets and returns the assignments from the RoadUserAssignmentRepository. If assignment creation is enabled and the repository is empty, the CreateRoadUserAssignments use case is triggered first.