reimplement filtering behavior of TrafficCounting
previously the event repository was filtered based on the interval defined by the counting specification, assignments were computed based on the filtered events

now assignments are computed on all events (leading to a more stable assignment, as assignments are now assigned to the same (longest) flow, independent of the applied counting interval filter)

the (optional) counting interval filter can be configured to have strict or open bounds: strict lower/upper bounds resp. require the start/end of the assignment to occur inside the interval.